refactor: Organize Pokémon card components.
This commit refactors the `MyPokemonCard` composable and its related UI elements into a more organized structure.

Key changes:
- **Component Refactoring:**
    - The `MyPokemonCard.kt` file has been cleaned up by extracting its internal composable functions (layers like `CapaFondo`, `CapaMarco`, `CapaUIDatos`, etc.) into their own files under a new `components` package (`app/src/main/java/com/amrubio27/mypokemoncard/pokemonCard/components/`).
    - `PokemonCardScreen.kt` and `EjemploPokemonCard.kt` are now top-level files in the `pokemonCard` package.
- **Drag Gesture Sensitivity Adjustment:**
    - The `factorSensibilidad` for drag gestures in `MyPokemonCard.kt` has been reduced from `0.4f` to `0.1f`, making the tilt effect less sensitive to drag movements.
- **Resource Updates:**
    - The application name in `strings.xml` has been changed from "MyPokemonCard" to "DevelopersTCG".
    - A new theme `Theme.MyPokemonCardWithActionBar` has been added to `themes.xml`, inheriting from `android:Theme.Material.Light.DarkActionBar`.
    - The `MainActivity` in `AndroidManifest.xml` now uses the new `Theme.MyPokemonCardWithActionBar` theme.
- **Code Structure and Imports:**
    - Imports in `MainActivity.kt` and `MyPokemonCard.kt` have been updated to reflect the new component locations.

This refactoring improves code organization and modularity, making it easier to manage.
