+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.MyPokemonCard">
+            android:theme="@style/Theme.MyPokemonCardWithActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
