+++ app/src/main/java/com/amrubio27/mypokemoncard/MainActivity.kt
@@ -4,14 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.amrubio27.mypokemoncard.pokemonCard.view.EjemploPokemonCard
+import com.amrubio27.mypokemoncard.pokemonCard.EjemploPokemonCard
 import com.amrubio27.mypokemoncard.ui.theme.MyPokemonCardTheme
 
 class MainActivity : ComponentActivity() {
